Backend/requirements.txt
Django==5.0
djangorestframework==3.14.0
mysqlclient==2.2.1
django-cors-headers==4.3.1
PyJWT==2.8.0
python-dotenv==1.0.0
django-filter==23.5
djangorestframework-simplejwt==5.3.1
